Add option to no. of chars in Tabs,some lli to li conversion thigies

+++ dcpp/DownloadManager.cpp
@@ -204,7 +204,7 @@ void DownloadManager::checkDownloads(UserConnection* aConn) {
 	}
 	fire(DownloadManagerListener::Requesting(), d);
 
-	dcdebug("Requesting " I64_FMT "/" I64_FMT "\n", d->getStartPos(), d->getSize());
+	dcdebug("Requesting " I64_FMT "/" I64_FMT "\n", (long long int)d->getStartPos(), d->getSize());
 
 	/*
 	find mySID, better ways to get the correct one transferred here?
@@ -249,7 +249,7 @@ void DownloadManager::startData(UserConnection* aSource, int64_t start, int64_t 
 	Download* d = aSource->getDownload();
 	dcassert(d != NULL);
 
-	dcdebug("Preparing " I64_FMT ":" I64_FMT ", " I64_FMT ":" I64_FMT"\n", d->getStartPos(), start, d->getSize(), bytes);
+	dcdebug("Preparing " I64_FMT ":" I64_FMT ", " I64_FMT ":" I64_FMT"\n", (long long int)d->getStartPos(), (long long int )start, (long long int)d->getSize(), (long long int)bytes);
 	if(d->getSize() == -1) {
 		if(bytes >= 0) {
 			d->setSize(bytes);
@@ -369,7 +369,7 @@ void DownloadManager::endData(UserConnection* aSource) {
 		aSource->setSpeed(d->getAverageSpeed());
 		aSource->updateChunkSize(d->getTigerTree().getBlockSize(), d->getSize(), GET_TICK() - d->getStart());
 
-		dcdebug("Download finished: %s, size " I64_FMT ", downloaded " I64_FMT "\n", d->getPath().c_str(), d->getSize(), d->getPos());
+		dcdebug("Download finished: %s, size " I64_FMT ", downloaded " I64_FMT "\n", d->getPath().c_str(), (long long int)d->getSize(), (long long int)d->getPos());
 
 		if(SETTING(LOG_DOWNLOADS) && (SETTING(LOG_FILELIST_TRANSFERS) || d->getType() == Transfer::TYPE_FILE) ) {
 			logDownload(aSource, d);
